Move base64_clip.py to a new devs/ folder
b64d/b64e are dev-tool functionality rather than general Windows
utilities, so they get their own top-level folder instead of living
in wins/ alongside clipsave.py.
